Allow selection of the PEAR DB driver to be used.
This allows using Aquarius on hosts where the mysqli extension is missing. Instead,
  $config['db']['driver'] = 'mysql'; // THIS IS NOT RECOMMENDED
can be used. This line is best put into config.local.php.

Of course you are free to try 'mssql', have fun swearing. It is our intention
to get 'sqlite' and 'pgsql' working, but this may happen after the sun has gone
nova, or not at all.