Prevent smarty from caching the Dynform rendering result.
I was pretty sure I'd confined the bits that turn on smarty caching to the frontend code. (Because this is the only place that needs it.) So this fix would not be necessary. Yet here I see a recent system building a smarty instance with caching already turned on. Has that commit fallen under the table in the last year or what? To be investigated.