Config option to ignore empty legends when saving
This is an ugly hack neccessary because some servers drop requests as they please. This may lead to deleted legends and consequently to irritated users.
With this hack enabled, legends cannot be emptied anymore. Not nice, but when your website dwells on an unreliable server this is a small issue.