Make archiver module more user-friendly.
Always deactivate the node once the archiving date is reached
unless the node is moved.

Historically, the archiving field required you to set sup1=1
so it would deactivate nodes. Otherwise, it would do nothing.
Yet in most setups, sup1 was not set initially because people
reasonably expected the field to work just like that.
To avoid future mistakes, if the node is not moved to another
place, and an archiving date is given, we just deactivate it
as a default.

The only way where this will have ill effects is in places
where the archiving field is defunct and serves as normal
date field. Suddenly, a slew of nodes will be deactivated.
It will suffice to just change those date fields to a
normal date field an check the log for all newly deactivated
nodes.